backend/requirements.txt
fastapi[all]
python-jose[cryptography]
yapf
python-dotenv
pyvis
networkx
transformers
langchain
accelerate
bitsandbytes
sentencepiece
nebula3-python
# llama-index == 0.10.12
# llama-index-core == 0.10.12
llama-index
llama-index-core
llama-index-graph-stores-nebula
llama-index-llms-langchain
llama-index-llms-huggingface
llama-index-embeddings-huggingface
vllm
llama-index-llms-vllm